#71 feat: 필터링 로직 분리 및 테스트 완료
